+++ R/cons_shape.R
@@ -1,0 +1,9 @@
+################################################################################
+#
+# Method to create a constraint matrix for shape-constrained splines
+# Works with various type of splines found in R.
+#
+################################################################################
+
+#' @export
+cons_shape <- function(x, ...) UseMethod("cons_shape")